requirements.txt
Django==5.1.1
pillow==10.4.0
environs[django]==9.3.2
python-telegram-bot==13.5
